Add PageBackDrop to pages, updated imports and minor styliing refactors
