+++ src/app/signup/page.tsx
@@ -1,10 +1,16 @@
 import SignUpForm from "./components/SignUpForm"
-import AuthCard from "@/components/ui/AuthCard"
+import { AuthCard, MoonGlow, PageBackdrop, StarField } from "@/components"
 
 export default function SignUp() {
     return (
-        <AuthCard title="Join the Guild">
-            <SignUpForm />
-        </AuthCard>
+        <>
+            <PageBackdrop>
+                <StarField />
+                <MoonGlow />
+            </PageBackdrop>
+            <AuthCard title="Join the Guild">
+                <SignUpForm />
+            </AuthCard>
+        </>
     )
 }
